+++ server/CHANGELOG.md
@@ -5,6 +5,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog][keepachangelog],
 and this project adheres to [Semantic Versioning][semver].
 
+## [Unreleased]
+
+### Fixed:
+
+- don't skip lines containing jinja expressions. This was causing some missing options when jinja was used in option
+
 ## [0.3.0] - 2020-02-23
 
 ### Added:
@@ -54,3 +60,4 @@ and this project adheres to [Semantic Versioning][semver].
 [0.2.0]: https://github.com/perrinjerome/vscode-zc-buildout/compare/v0.1.1...v0.2.0
 [0.2.1]: https://github.com/perrinjerome/vscode-zc-buildout/compare/v0.2.0...v0.2.1
 [0.3.0]: https://github.com/perrinjerome/vscode-zc-buildout/compare/v0.2.1...v0.3.0
+[unreleased]: https://github.com/perrinjerome/vscode-zc-buildout/compare/v0.3.0...master